Concat Palindrome
Chef has two strings 
�
A and 
�
B of lengths 
�
N and 
�
M respectively.

Chef can rearrange both strings in any way he wants. Let the rearranged string of 
�
A be 
�
′
A 
′
  and the rearranged string of 
�
B be 
�
′
B 
′
 .

Determine whether we can construct rearrangements 
�
′
A 
′
  and 
�
′
B 
′
  such that 
(
�
′
+
�
′
)
(A 
′
 +B 
′
 ) is a palindrome.

Here 
+
+ denotes the concatenation operation. For e.g. 
�
�
�
+
�
�
�
=
�
�
�
�
�
�
abc+xyz=abcxyz.

Note: A string is called palindrome if it reads the same backwards and forwards, for e.g. 
noon
noon and 
level
level are palindromic strings.